requirements.txt
easyocr==1.7.2
beautifulsoup4==4.11.2
requests==2.28.1
python-dotenv==1.0.1
black==24.4.0